+++ qsb/lua/modules/QSB_3_BriefingSystem/api.lua
@@ -265,7 +265,7 @@ function API.StartBriefing(_Briefing, _Name, _PlayerID)
     end
     local PlayerID = _PlayerID;
     if not PlayerID and not Framework.IsNetworkGame() then
-        PlayerID = QSB.HumanPlayerID;
+        PlayerID = 1;
     end
     assert(_Name ~= nil);
     assert(_PlayerID ~= nil);