+++ qsb/lua/modules/QSB_3_DialogSystem/api.lua
@@ -122,7 +122,7 @@ function API.StartDialog(_Dialog, _Name, _PlayerID)
     end
     local PlayerID = _PlayerID;
     if not PlayerID and not Framework.IsNetworkGame() then
-        PlayerID = QSB.HumanPlayerID;
+        PlayerID = 1;
     end
     assert(_Name ~= nil);
     assert(_PlayerID ~= nil);